http: add option for slowing down receive of large http responses
This allows a client to receive data from a server in a controlled fashion.
Specifically, the client can choose to stop reading data from the server for
some amount of time.

It is still not possible for a server to receive a very large (larger than the
amount of memory available to the process) POST in a controlled fashion... this
would be a nice piece of functionality to add sometime.